+++ GUI/Misc/AchievementFrameTabButton.lua
@@ -19,8 +19,8 @@ function achFrameTabBtn:New(text, framesToShow, categories, filters, waterMark)
     addon.Util.InjectMetatable(frame, achFrameTabBtn);
 
 	-- Set properties
-    frame.ID = tabId;
-    tinsert(ourTabIDs, frame.ID);
+    frame.Id = tabId;
+    tinsert(ourTabIDs, frame.Id);
     frame.FramesToShow = framesToShow;
 
     frame.SelectedAchievement = nil; -- Issue #6: Fix
@@ -45,11 +45,11 @@ function achFrameTabBtn:New(text, framesToShow, categories, filters, waterMark)
 
     if addon.IsWrathClassic then
         hooksecurefunc("PanelTemplates_SetTab", function(_, clickedTab) -- Issue #1: Broken
-            frame:AchievementFrame_UpdateTabs(frame, frame.ID, clickedTab);
+            frame:AchievementFrame_UpdateTabs(frame, frame.Id, clickedTab);
         end);
     else
         hooksecurefunc("AchievementFrame_UpdateTabs", function(clickedTab) -- Issue #1: Broken
-            frame:AchievementFrame_UpdateTabs(frame, frame.ID, clickedTab);
+            frame:AchievementFrame_UpdateTabs(frame, frame.Id, clickedTab);
         end);
     end
 
@@ -64,12 +64,8 @@ function achFrameTabBtn:Base_OnClick(id)
         PanelTemplates_Tab_OnClick(_G["AchievementFrameTab" .. id], AchievementFrame);
     else
         if addon.InGuildView() then
-            if addon.IsShadowlandsRetail then
-                AchievementFrame_ToggleView();
-            else
-                AchievementFrameBaseTab_OnClick(1);
-                AchievementFrame_RefreshView();
-            end
+            AchievementFrameBaseTab_OnClick(1);
+            AchievementFrame_RefreshView();
             AchievementFrameGuildEmblemLeft:Hide();
             AchievementFrameGuildEmblemRight:Hide();
         end
@@ -139,6 +135,6 @@ end
 
 function achFrameTabBtn:Select()
     if gui.SelectedTab ~= self then
-        self:OnClick(self.ID);
+        self:OnClick(self.Id);
     end
 end