+++ Localization/deDE.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-01-16 17-36-03 ]] --
+-- [[ Exported at 2023-01-19 17-25-20 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -14,6 +14,7 @@ L["TAB"] = tab;
 
 -- Load strings into the localization that are already localized by Blizzard
 addon.ExtraWrathStrings.deDE.Load(L);
+addon.PluginStrings.deDE.Load(L);
 addon.GlobalStrings.Load(L);
 addon.Plugins:LoadLocalization(L);