+++ Localization/enUS.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-01-16 17-36-03 ]] --
+-- [[ Exported at 2023-01-19 17-25-20 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --