+++ Localization/zhCN.PluginStrings.lua
@@ -1,0 +1,13 @@
+-- [[ Exported at 2023-01-19 17-25-20 ]] --
+-- [[ This code is automatically generated as an export from ]] --
+-- [[ an SQLite database and is not meant for manual edit. ]] --
+
+-- [[ Namespaces ]] --
+local _, addon = ...;
+addon.PluginStrings.zhCN = {};
+local pluginStrings = addon.PluginStrings.zhCN;
+
+function pluginStrings.Load(L)
+
+end
+